Fix security extension split package issue #1060
