+++ extensions/security/runtime/src/main/java/io/quarkus/security/runtime/SecurityConfig.java
@@ -1,4 +1,4 @@
-package io.quarkus.security;
+package io.quarkus.security.runtime;
 
 import io.quarkus.runtime.annotations.ConfigItem;
 import io.quarkus.runtime.annotations.ConfigPhase;
@@ -8,15 +8,15 @@ import io.quarkus.runtime.annotations.ConfigRoot;
  *
  */
 @ConfigRoot(phase = ConfigPhase.RUN_TIME)
-final class SecurityConfig {
+public final class SecurityConfig {
     /**
      * The configuration for the {@linkplain org.wildfly.security.auth.realm.LegacyPropertiesSecurityRealm}
      */
     @ConfigItem
-    PropertiesRealmConfig file;
+    public PropertiesRealmConfig file;
     /**
      * The configuration for the {@linkplain org.wildfly.security.auth.realm.SimpleMapBackedSecurityRealm}
      */
     @ConfigItem
-    MPRealmConfig embedded;
+    public MPRealmConfig embedded;
 }
