+++ extensions/smallrye-jwt/deployment/src/main/java/io/quarkus/smallrye/jwt/deployment/SmallRyeJwtProcessor.java
@@ -31,11 +31,11 @@ import io.quarkus.deployment.builditem.FeatureBuildItem;
 import io.quarkus.deployment.builditem.ObjectSubstitutionBuildItem;
 import io.quarkus.deployment.builditem.substrate.ReflectiveClassBuildItem;
 import io.quarkus.runtime.RuntimeValue;
-import io.quarkus.security.AuthConfig;
 import io.quarkus.security.AuthConfigBuildItem;
 import io.quarkus.security.IdentityManagerBuildItem;
 import io.quarkus.security.SecurityDomainBuildItem;
 import io.quarkus.security.SecurityRealmBuildItem;
+import io.quarkus.security.runtime.AuthConfig;
 import io.quarkus.smallrye.jwt.runtime.ClaimValueProducer;
 import io.quarkus.smallrye.jwt.runtime.CommonJwtProducer;
 import io.quarkus.smallrye.jwt.runtime.JWTAuthContextInfoGroup;
